Move NowcastData into data pipeline
