+++ pipelines/data/hubverse_nowcast.py
@@ -9,9 +9,10 @@ from typing import TYPE_CHECKING
 
 import polars as pl
 
+from pipelines.data.nowcast import NowcastData
+
 if TYPE_CHECKING:
     from pipelines.epiautogp.forecast_spec import ForecastSpec
-    from pipelines.epiautogp.nowcast import NowcastData
 
 HUBVERSE_MODEL_OUTPUT_SUBDIR = Path("model-output", "CFA-nowcastNHSN")
 HUBVERSE_TARGETS = {
@@ -186,7 +187,5 @@ class HubverseNowcast:
                 "output_type_id", maintain_order=True
             )
         ]
-
-        from pipelines.epiautogp.nowcast import NowcastData
 
         return NowcastData(dates=nowcast_dates, reports=nowcast_reports)
