+++ pipelines/epiautogp/README.md
@@ -69,6 +69,7 @@ Shared utilities for the forecast pipeline, containing modular functions for eac
 
 Pluggable nowcasting sources for nowcasting recent observations.
 
+- **`NowcastData`** (`../data/nowcast.py`): Stores nowcast dates and report trajectories.
 - **`NowcastSource`** (Protocol): Declares `applies_to(*, forecast_spec) -> bool` (the predicate the resolver queries before constructing) and `get_nowcast_data(*, dates, reports) -> NowcastData` (the action).
 - **`FixedNowcast`**: Trivial source wrapping a precomputed `NowcastData`.
 - **`ReportingDelayNowcast`**: Inflates the most-recent observations by the inverse of a reporting-delay PMF.
