+++ pipelines/epiautogp/__init__.py
@@ -3,8 +3,9 @@ EpiAutoGP integration module for cfa-stf-routine-forecasting pipelines.
 """
 
 from pipelines.data.hubverse_nowcast import HubverseNowcast
+from pipelines.data.nowcast import NowcastData
 from pipelines.epiautogp.epiautogp_forecast_utils import setup_forecast_pipeline
-from pipelines.epiautogp.nowcast import FixedNowcast, NowcastData, NowcastSource
+from pipelines.epiautogp.nowcast import FixedNowcast, NowcastSource
 from pipelines.epiautogp.prep_epiautogp_data import convert_to_epiautogp_json
 from pipelines.epiautogp.reporting_delay import (
     inflate_report,
